examples/dj/requirements.txt
Django==1.7.1
argparse==1.2.1
coverage==3.7.1
django-nose==1.3
django-whatever==0.2.3
djangorestframework==3.0.0
nose==1.3.4
wsgiref==0.1.2
